fix: move user4 test to flaky
