+++ test/source/tests/flaky.ts
@@ -19,6 +19,7 @@ import { SetupPageRecipe } from './page-recipe/setup-page-recipe';
 import { KeyUtil } from '../core/crypto/key';
 import { ElementHandle } from 'puppeteer';
 import { expectRecipientElements } from './compose';
+import { GoogleData } from '../mock/google/google-data';
 
 // tslint:disable:no-blank-lines-func
 
@@ -140,6 +141,89 @@ export const defineFlakyTests = (testVariant: TestVariant, testWithBrowser: Test
       expect(key.algo.algorithm).to.equal('rsa_encrypt_sign');
       await myKeyFrame.close();
       await settingsPage.close();
+    }));
+
+    /**
+     * You need the following lines in /etc/hosts:
+     * 127.0.0.1    standardsubdomainfes.test
+     * 127.0.0.1    fes.standardsubdomainfes.test
+     */
+    ava.default('user4@standardsubdomainfes.test:8001 - PWD encrypted message with FES web portal - some sends fail with BadRequest error', testWithBrowser(undefined, async (t, browser) => {
+      const acct = 'user4@standardsubdomainfes.test:8001'; // added port to trick extension into calling the mock
+      const settingsPage = await BrowserRecipe.openSettingsLoginApprove(t, browser, acct);
+      await SetupPageRecipe.manualEnter(settingsPage, 'flowcrypt.test.key.used.pgp', { submitPubkey: false, usedPgpBefore: false },
+        { isSavePassphraseChecked: false, isSavePassphraseHidden: false });
+      // add a name to one of the contacts
+      const dbPage = await browser.newPage(t, TestUrls.extension('chrome/dev/ci_unit_test.htm'));
+      await dbPage.page.evaluate(async () => {
+        const db = await (window as any).ContactStore.dbOpen();
+        await (window as any).ContactStore.update(db, 'cc@example.com', { name: 'Mr Cc' });
+      });
+      await dbPage.close();
+      const subject = 'PWD encrypted message with FES web portal - some sends fail with BadRequest error - ' + testVariant;
+      let expectedNumberOfPassedMessages = (await GoogleData.withInitializedData(acct)).searchMessagesBySubject(subject).length;
+      // 1. vague Gmail error with partial success
+      let composePage = await ComposePageRecipe.openStandalone(t, browser, 'user4@standardsubdomainfes.test:8001');
+      await ComposePageRecipe.fillMsg(composePage, { to: 'to@example.com', cc: 'cc@example.com', bcc: 'flowcrypt.compatibility@gmail.com' }, subject);
+      await composePage.waitAndType('@input-password', 'gO0d-pwd');
+      await composePage.waitAndClick('@action-send', { delay: 1 });
+      await composePage.waitAndRespondToModal('confirm', 'cancel',
+        'Messages to some recipients were sent successfully, while messages to flowcrypt.compatibility@gmail.com, Mr Cc <cc@example.com> ' +
+        'encountered error(s) from Gmail. Please help us improve FlowCrypt by reporting the error to us.');
+      await composePage.close();
+      expect((await GoogleData.withInitializedData(acct)).searchMessagesBySubject(subject).length).to.equal(++expectedNumberOfPassedMessages);
+      // 2. vague Gmail error with all failures
+      composePage = await ComposePageRecipe.openStandalone(t, browser, 'user4@standardsubdomainfes.test:8001');
+      await ComposePageRecipe.fillMsg(composePage, { cc: 'cc@example.com', bcc: 'flowcrypt.compatibility@gmail.com' }, subject);
+      await composePage.waitAndType('@input-password', 'gO0d-pwd');
+      await composePage.waitAndClick('@action-send', { delay: 1 });
+      await composePage.waitAndRespondToModal('confirm', 'cancel',
+        'Google returned an error when sending message. ' +
+        'Please help us improve FlowCrypt by reporting the error to us.');
+      await composePage.close();
+      expect((await GoogleData.withInitializedData(acct)).searchMessagesBySubject(subject).length).to.equal(expectedNumberOfPassedMessages); // + 0 messages
+      // 3. "invalid To" Gmail error with partial success
+      composePage = await ComposePageRecipe.openStandalone(t, browser, 'user4@standardsubdomainfes.test:8001');
+      await ComposePageRecipe.fillMsg(composePage, { to: 'invalid@example.com', cc: 'to@example.com' }, subject);
+      await composePage.waitAndType('@input-password', 'gO0d-pwd');
+      await composePage.waitAndClick('@action-send', { delay: 1 });
+      await composePage.waitAndRespondToModal('error', 'confirm',
+        'Messages to some recipients were sent successfully, while messages to invalid@example.com ' +
+        'encountered error(s) from Gmail: Invalid recipients\n\nPlease remove recipients, add them back and re-send the message.');
+      await composePage.close();
+      expect((await GoogleData.withInitializedData(acct)).searchMessagesBySubject(subject).length).to.equal(++expectedNumberOfPassedMessages);
+      // 4. "invalid To" Gmail error with all failures
+      composePage = await ComposePageRecipe.openStandalone(t, browser, 'user4@standardsubdomainfes.test:8001');
+      await ComposePageRecipe.fillMsg(composePage, { to: 'invalid@example.com', cc: 'cc@example.com' }, subject);
+      await composePage.waitAndType('@input-password', 'gO0d-pwd');
+      await composePage.waitAndClick('@action-send', { delay: 1 });
+      await composePage.waitAndRespondToModal('error', 'confirm',
+        'Error from google: Invalid recipients\n\nPlease remove recipients, add them back and re-send the message.');
+      await composePage.close();
+      expect((await GoogleData.withInitializedData(acct)).searchMessagesBySubject(subject).length).to.equal(expectedNumberOfPassedMessages); // + 0 messages
+      // 5. "RequestTimeout" error with partial success
+      composePage = await ComposePageRecipe.openStandalone(t, browser, 'user4@standardsubdomainfes.test:8001');
+      await ComposePageRecipe.fillMsg(composePage, { to: 'timeout@example.com', cc: 'to@example.com' }, subject);
+      await composePage.waitAndType('@input-password', 'gO0d-pwd');
+      await composePage.waitAndClick('@action-send', { delay: 1 });
+      await composePage.waitAndRespondToModal('error', 'confirm',
+        'Messages to some recipients were sent successfully, while messages to timeout@example.com ' +
+        'encountered network errors. Please check your internet connection and try again.');
+      await composePage.close();
+      expect((await GoogleData.withInitializedData(acct)).searchMessagesBySubject(subject).length).to.equal(++expectedNumberOfPassedMessages);
+      // 6. "RequestTimeout" error with all failures
+      composePage = await ComposePageRecipe.openStandalone(t, browser, 'user4@standardsubdomainfes.test:8001');
+      await ComposePageRecipe.fillMsg(composePage, { to: 'timeout@example.com', cc: 'cc@example.com' }, subject);
+      await composePage.waitAndType('@input-password', 'gO0d-pwd');
+      await composePage.waitAndClick('@action-send', { delay: 1 });
+      await composePage.waitAndRespondToModal('error', 'confirm',
+        'Could not send message due to network error. Please check your internet connection and try again. ' +
+        '(This may also be caused by missing extension permissions).');
+      await composePage.close();
+      expect((await GoogleData.withInitializedData(acct)).searchMessagesBySubject(subject).length).to.equal(expectedNumberOfPassedMessages); // + 0 messages
+      // this test is using PwdEncryptedMessageWithFesReplyBadRequestTestStrategy to check sent result based on subject
+      // "PWD encrypted message with FES web portal - some sends fail with BadRequest error"
+      // also see '/api/v1/message' in fes-endpoints.ts mock
     }));
 
     ava.default('user@forbid-storing-passphrase-client-configuration.flowcrypt.test - do not store passphrase', testWithBrowser(undefined, async (t, browser) => {
